Working commit readiness
Moved out API fully from sensors
Fixed token handling so only one token per sensor group
