+++ custom_components/oura/views.py
@@ -3,6 +3,9 @@
 from homeassistant import core
 from homeassistant.components import http
 import json
+import logging
+
+_LOGGER = logging.getLogger(__name__)
 
 # Views configuration.
 AUTH_CALLBACK_NAME = 'api:oura'
@@ -20,13 +23,13 @@ class OuraAuthCallbackView(http.HomeAssistantView):
   url = AUTH_CALLBACK_PATH
   name = AUTH_CALLBACK_NAME
 
-  def __init__(self, sensor):
+  def __init__(self, api):
     """Initializes view.
 
     Args:
       sensor: Sensor which initialized the OAuth process.
     """
-    self._sensor = sensor
+    self._api = api
 
   @core.callback
   def get(self, request):
@@ -39,12 +42,14 @@ class OuraAuthCallbackView(http.HomeAssistantView):
     code_data = {'code': code}
 
     sensor_name = request.query.get('state')
-    token_file_name = self._sensor._api.token_file_name
+    token_file_name = self._api.token_file_name
+    _LOGGER.info("token_file_name %s.", token_file_name) 
 
     with open(token_file_name, 'w+') as token_file:
+      _LOGGER.info("writing to token file") 
       token_file.write(json.dumps(code_data))
 
-    self._sensor.update()  # Forces exchanging code for access token.
+    self._api._get_access_token_data_from_file()
 
     return self.json_message(
         f'Oura OAuth code {code} for sensor.{sensor_name} stored in '
